feat: support undo/redo in set bold
