+++ packages/base-docs/src/Basics/commonParameter.ts
@@ -1,0 +1,14 @@
+class CommonParameter {
+    cursor: number = 0;
+
+    reset() {
+        this.cursor = 0;
+        return this;
+    }
+
+    moveCursor(pos: number) {
+        this.cursor += pos;
+    }
+}
+
+export default CommonParameter;
